Split the Plugin template into multiple files
