+++ src/components/backstage/plugins/Changelog.js
@@ -1,0 +1,53 @@
+import React from 'react';
+import isEmpty from 'lodash/isEmpty';
+import { Link, Title } from 'components';
+
+import PAGE_SECTIONS from './pageSections';
+
+const Changelog = ({ plugin }) => {
+  const { changelogSince, changelog, humanName, codeLocation } = plugin;
+  if (isEmpty(plugin[PAGE_SECTIONS.CHANGELOG.existsKey])) return null;
+
+  let inner = (
+    <div className="prose prose-primary max-w-none mb-4">
+      <p>
+        The {humanName} plugin has not seen any significant changes since {changelogSince}.
+      </p>
+    </div>
+  );
+
+  if (!isEmpty(changelog)) {
+    inner = (
+      <>
+        <div className="prose prose-primary max-w-none mb-4">
+          <p>
+            This changelog is produced from commits made to the {humanName} plugin since{' '}
+            {changelogSince}, and based on the code located{' '}
+            <Link to={codeLocation} color="primary">
+              here
+            </Link>
+            . It may not contain information about all commits. Releases and version bumps are
+            intentionally omitted. This changelog is generated by AI.
+          </p>
+        </div>
+
+        <div
+          className="prose prose-primary max-w-none"
+          dangerouslySetInnerHTML={{ __html: changelog.childMarkdownRemark.html }}
+        />
+      </>
+    );
+  }
+
+  return (
+    <div className="mb-24">
+      <div className="mb-4" id={PAGE_SECTIONS.CHANGELOG.fragment}>
+        <Title className="text-3xl">{PAGE_SECTIONS.CHANGELOG.label}</Title>
+      </div>
+
+      {inner}
+    </div>
+  );
+};
+
+export default Changelog;
