+++ src/components/backstage/plugins/index.js
@@ -8,3 +8,5 @@ export { default as RoadieDocsChip } from './RoadieDocsChip';
 export { default as GitHubChip } from './GitHubChip';
 export { default as NpmChip } from './NpmChip';
 export { default as GravatarImage } from './GravatarImage';
+export { default as Body } from './Body';
+export { default as PAGE_SECTIONS } from './pageSections';
