+++ src/components/backstage/plugins/pageSections.js
@@ -1,0 +1,29 @@
+
+const PAGE_SECTIONS = {
+  INTRODUCTION: {
+    key: 'introduction',
+    existsKey: 'introduction',
+    fragment: 'introduction',
+    label: 'Introduction',
+  },
+  INSTALLATION_INSTRUCTIONS: {
+    key: 'installationInstructions',
+    existsKey: 'installationInstructions',
+    fragment: 'installation-instructions',
+    label: 'Installation Instructions',
+  },
+  NOTES: {
+    key: 'notes',
+    existsKey: 'notes',
+    fragment: 'notes',
+    label: 'Things to Know',
+  },
+  CHANGELOG: {
+    key: 'changelog',
+    existsKey: 'changelogSince',
+    fragment: 'changelog',
+    label: 'Changelog',
+  },
+};
+
+export default PAGE_SECTIONS;
